Release v.0.0.2 Contains the following enhancements: - Adds feature to move textfields/textview contained in a cell above keyboards. - Also adds in other tweaks for dismissal and a few bug fixes. - New delegate for when last item is dismissed.

+++ HTKScrollingNavigationController Demo/HTKScrollingNavigation/HTKSampleViewController.m
@@ -37,7 +37,7 @@
     // randomly size and determine background color
     NSArray *bgColorArray = @[[UIColor redColor], [UIColor purpleColor], [UIColor orangeColor], [UIColor greenColor], [UIColor yellowColor], [UIColor whiteColor], [UIColor cyanColor], [UIColor magentaColor], [UIColor brownColor]];
     // set random color
-    self.view.backgroundColor = bgColorArray[arc4random_uniform(bgColorArray.count)];
+    self.view.backgroundColor = bgColorArray[arc4random_uniform((int)bgColorArray.count)];
     
     // add two buttons one for push, one for pop
     UIButton *pushButton = [UIButton buttonWithType:UIButtonTypeCustom];
